+++ dramatiq/common.py
@@ -14,7 +14,6 @@
 #
 # You should have received a copy of the GNU Lesser General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
 from queue import Empty
 from random import uniform
 from time import time